Add max frame time to all tests and move debug it config.
