+++ packages/kartoffelgames.web.potato_web_builder/source/core/core_entity/core-entity-updater.ts
@@ -1,6 +1,6 @@
 import { Stack } from '@kartoffelgames/core';
 import { InteractionEvent, InteractionZone } from '@kartoffelgames/web.interaction-zone';
-import { PwbDebug, PwbDebugLogLevel } from '../../debug/pwb-debug';
+import { PwbDebug, PwbDebugLogLevel } from '../configuration/pwb-debug';
 import { UpdateTrigger } from '../enum/update-trigger.enum';
 import { CoreEntityInteractionData, CoreEntityInteractionEvent, CoreEntityProcessorProxy } from './interaction-tracker/core-entity-processor-proxy';
 import { IgnoreInteractionTracking } from './interaction-tracker/ignore-interaction-tracking.decorator';
