+++ packages/kartoffelgames.web.potato_web_builder/source/core/module/base-module.ts
@@ -3,7 +3,7 @@ import { ScopedValues } from '../scoped-values';
 import { CoreEntityProcessorConstructor } from '../core_entity/core-entity';
 import { CoreEntityExtendable, CoreEntityExtendableConstructorParameter } from '../core_entity/core-entity-extendable';
 import { ModuleValues } from './module-values';
-import { PwbDebugLogLevel } from '../../debug/pwb-debug';
+import { PwbDebugLogLevel } from '../configuration/pwb-debug';
 import { Processor } from '../core_entity/processor';
 
 export abstract class BaseModule<TModuleProcessor extends IPwbModuleProcessor> extends CoreEntityExtendable<TModuleProcessor> implements IDeconstructable {
