+++ packages/kartoffelgames.web.potato_web_builder/test/integration/default/event-listener.test.ts
@@ -13,6 +13,7 @@ import { InstructionResult } from '../../../source/core/module/instruction_modul
 import { PwbInstructionModule } from '../../../source/core/module/instruction_module/pwb-instruction-module.decorator';
 import { ModuleValues } from '../../../source/core/module/module-values';
 import { ScopedValues } from '../../../source/core/scoped-values';
+import { PwbDebug } from '../../../source/core/configuration/pwb-debug';
 import { PwbComponentEventListener } from '../../../source/module/component-event-listener/pwb-component-event-listener.decorator';
 import { ComponentEvent } from '../../../source/module/component-event/component-event';
 import { ComponentEventEmitter } from '../../../source/module/component-event/component-event-emitter';
@@ -23,6 +24,11 @@ import '../../utility/chai-helper';
 import { TestUtil } from '../../utility/test-util';
 
 describe('PwbEventListener', () => {
+    before(() => {
+        const lConfiguration: PwbDebug = new PwbDebug();
+        lConfiguration.configuration.updating.frameTime = Number.MAX_SAFE_INTEGER;
+    });
+
     it('-- Native listener', async () => {
         // Process.
         let lEventCalled: boolean = false;
