+++ packages/kartoffelgames.web.potato_web_builder/test/integration/extension/component-event-emitter-extension.test.ts
@@ -2,6 +2,7 @@ import { Exception } from '@kartoffelgames/core';
 import { expect } from 'chai';
 import { Processor } from '../../../source';
 import { PwbComponent } from '../../../source/core/component/pwb-component.decorator';
+import { PwbDebug } from '../../../source/core/configuration/pwb-debug';
 import { ComponentEvent } from '../../../source/module/component-event/component-event';
 import { ComponentEventEmitter } from '../../../source/module/component-event/component-event-emitter';
 import { PwbComponentEvent } from '../../../source/module/component-event/pwb-component-event.decorator';
@@ -11,6 +12,11 @@ import '../../utility/chai-helper';
 import { TestUtil } from '../../utility/test-util';
 
 describe('ComponentEventEmitterExtension', () => {
+    before(() => {
+        const lConfiguration: PwbDebug = new PwbDebug();
+        lConfiguration.configuration.updating.frameTime = Number.MAX_SAFE_INTEGER;
+    });
+
     it('-- Correct event value', async () => {
         // Setup. Values.
         const lEventValue: string = 'EVENT-VALUE';
