+++ packages/kartoffelgames.web.potato_web_builder/test/integration/module/event-attribute.module.test.ts
@@ -1,11 +1,17 @@
 import { expect } from 'chai';
 import { PwbComponent } from '../../../source/core/component/pwb-component.decorator';
 import { Processor } from '../../../source/core/core_entity/processor';
+import { PwbDebug } from '../../../source/core/configuration/pwb-debug';
 import '../../mock/request-animation-frame-mock-session';
 import '../../utility/chai-helper';
 import { TestUtil } from '../../utility/test-util';
 
 describe('EventAttributeModule', () => {
+    before(() => {
+        const lConfiguration: PwbDebug = new PwbDebug();
+        lConfiguration.configuration.updating.frameTime = Number.MAX_SAFE_INTEGER;
+    });
+
     it('-- Basic click event', async () => {
         // Setup. Values.
         const lEventComponentSelector: string = TestUtil.randomSelector();
