+++ packages/kartoffelgames.web.potato_web_builder/test/integration/module/pwb-child-attribute.module.test.ts
@@ -3,6 +3,7 @@ import { expect } from 'chai';
 import { PwbComponent } from '../../../source/core/component/pwb-component.decorator';
 import { CoreEntityProcessorProxy } from '../../../source/core/core_entity/interaction-tracker/core-entity-processor-proxy';
 import { Processor } from '../../../source/core/core_entity/processor';
+import { PwbDebug } from '../../../source/core/configuration/pwb-debug';
 import { PwbExport } from '../../../source/module/export/pwb-export.decorator';
 import { PwbChild } from '../../../source/module/pwb_child/pwb-child.decorator';
 import '../../mock/request-animation-frame-mock-session';
@@ -10,6 +11,11 @@ import '../../utility/chai-helper';
 import { TestUtil } from '../../utility/test-util';
 
 describe('PwbChildAttributeModule', () => {
+    before(() => {
+        const lConfiguration: PwbDebug = new PwbDebug();
+        lConfiguration.configuration.updating.frameTime = Number.MAX_SAFE_INTEGER;
+    });
+
     it('-- Read id child', async () => {
         // Setup. Values.
         const lIdName: string = 'IdChildId';
