+++ packages/kartoffelgames.web.potato_web_builder/test/integration/module/slot-instruction-module.test.ts
@@ -2,6 +2,7 @@ import { InjectionConstructor } from '@kartoffelgames/core.dependency-injection'
 import { expect } from 'chai';
 import { PwbComponent } from '../../../source/core/component/pwb-component.decorator';
 import { Processor } from '../../../source/core/core_entity/processor';
+import { PwbDebug } from '../../../source/core/configuration/pwb-debug';
 import '../../mock/request-animation-frame-mock-session';
 import '../../utility/chai-helper';
 import { TestUtil } from '../../utility/test-util';
@@ -10,6 +11,11 @@ import { TestUtil } from '../../utility/test-util';
 const HTMLSlotElement: InjectionConstructor = <any>document.createElement('slot').constructor;
 
 describe('SlotAttribute', () => {
+    before(() => {
+        const lConfiguration: PwbDebug = new PwbDebug();
+        lConfiguration.configuration.updating.frameTime = Number.MAX_SAFE_INTEGER;
+    });
+
     it('-- Default slot', async () => {
         // Setup. Define component.
         @PwbComponent({
